feat(edit): working tile update
