+++ next-tavla/app/(admin)/edit/[id]/actions.ts
@@ -18,11 +18,3 @@ export async function addTile(bid: TBoardID, tile: TTile) {
         .update({ tiles: firestore.FieldValue.arrayUnion(tile) })
 }
 
-export async function saveTile(bid: TBoardID, tile: TTile) {
-    const docRef = firestore().collection('boards').doc(bid)
-    const doc = (await docRef.get()).data() as TBoard
-    const oldTile = doc.tiles.find((t) => t.uuid === tile.uuid)
-    if (oldTile)
-        docRef.update({ tiles: firestore.FieldValue.arrayRemove(oldTile) })
-    docRef.update({ tiles: firestore.FieldValue.arrayUnion(tile) })
-}
