+++ next-tavla/app/(admin)/edit/[id]/components/TileCard/actions.ts
@@ -1,7 +1,7 @@
 'use server'
 import { firestore } from 'firebase-admin'
 import { initializeAdminApp } from 'Admin/utils/firebase'
-import { TBoardID } from 'types/settings'
+import { TBoard, TBoardID } from 'types/settings'
 import { TTile } from 'types/tile'
 import { revalidatePath } from 'next/cache'
 
@@ -14,3 +14,14 @@ export async function deleteTile(bid: TBoardID, tile: TTile) {
         .update({ tiles: firestore.FieldValue.arrayRemove(tile) })
     revalidatePath(`/edit/${bid}`)
 }
+
+export async function saveTile(bid: TBoardID, tile: TTile) {
+    const docRef = firestore().collection('boards').doc(bid)
+    const doc = (await docRef.get()).data() as TBoard
+    const oldTile = doc.tiles.find((t) => t.uuid === tile.uuid)
+    if (oldTile)
+        docRef.update({ tiles: firestore.FieldValue.arrayRemove(oldTile) })
+    docRef.update({ tiles: firestore.FieldValue.arrayUnion(tile) })
+    revalidatePath(`/edit/${bid}`)
+}
+
